Enable run_retries by default
Summary:
The default maximum number of retries is still set to 0, but there is longer a need to enable the feature as well. This allows the config instructions to be the same between OSS and cloud. It also matches the current defaults for the helm chart

Test Plan: BK
run dagster dev, add a retry tag to a job, make it fail, watch it retry